Update pdb-addr2line and pdb with it.
The only change in pdb-addr2line 0.9 compared to 0.7 was the update of
the pdb crate from 0.7 to 0.8. Now module index is a usize, and type
sizes are u64.

This update de-duplicates our uuid and pdb dependencies.

+++ src/windows/symbol.rs
@@ -177,7 +177,7 @@ impl SelectedSymbol {
         } else {
             match formatter.format_function(
                 &self.name,
-                self.module_index.unwrap_or(0) as u16,
+                self.module_index.unwrap_or(0),
                 self.type_index,
             ) {
                 Ok(function) => FuncName::Undecorated(function),
@@ -210,12 +210,12 @@ impl SelectedSymbol {
             return self.parameter_size;
         }
 
-        let module_index = self.module_index.unwrap_or(0) as u16;
+        let module_index = self.module_index.unwrap_or(0);
         let (min_start, max_end) = self.ebp.drain(..).fold((std::u32::MAX, 0), |acc, i| {
             (
                 acc.0.min(i.offset),
                 acc.1
-                    .max(i.offset + formatter.get_type_size(module_index, i.type_index)),
+                    .max(i.offset + formatter.get_type_size(module_index, i.type_index) as u32),
             )
         });
